[IE Myriad] Extract MyriadDeviceProvider class from MyriadExecutor. (Without refactoring existing methods)
