+++ inference-engine/src/vpu/myriad_plugin/myriad_executor.h
@@ -13,6 +13,7 @@
 
 #include <mvnc.h>
 #include "myriad_mvnc_wraper.h"
+#include "myriad_device_provider.h"
 
 #include <ie_parameter.hpp>
 
@@ -32,63 +33,13 @@ struct GraphDesc {
     ncFifoHandle_t *_outputFifoHandle = nullptr;
 };
 
-struct DeviceDesc {
-    int _graphNum = 0;
-    int _maxGraphNum = 0;
-    std::string _name;
-    ncDevicePlatform_t _platform = NC_ANY_PLATFORM;
-    ncDeviceProtocol_t _protocol = NC_ANY_PROTOCOL;
-
-    int _deviceIdx = -1;
-    ncDeviceHandle_t *_deviceHandle = nullptr;
-
-    bool isBooted() const {
-        return _deviceHandle != nullptr;
-    }
-    bool isEmpty() const {
-        return _graphNum == 0;
-    }
-    bool isNotFull() const {
-        return _graphNum < _maxGraphNum;
-    }
-
-    bool isSuitableForConfig(const MyriadConfig& config) const {
-        bool isSuitableByName = true;
-        if (!config.deviceName().empty()) {
-            isSuitableByName = config.deviceName() == _name;
-        }
-
-        return isSuitableByName &&
-                ((config.platform() == NC_ANY_PLATFORM) || (_platform == config.platform())) &&
-                ((config.protocol() == NC_ANY_PROTOCOL) || (_protocol == config.protocol()));
-    }
-
-    Platform revision() const {
-        VPU_THROW_UNLESS(_platform != NC_ANY_PLATFORM, "Cannot get a revision from not booted device");
-        return _platform == NC_MYRIAD_2 ? Platform::MYRIAD_2 : Platform::MYRIAD_X;
-    }
-};
-
-typedef std::shared_ptr<DeviceDesc> DevicePtr;
-
-
 class MyriadExecutor {
     Logger::Ptr _log;
-    std::shared_ptr<IMvnc> _mvnc;
     unsigned int _numStages = 0;
 
 public:
-    MyriadExecutor(bool forceReset, std::shared_ptr<IMvnc> mvnc,
-                        const LogLevel& vpuLogLevel, const Logger::Ptr& log);
+    explicit MyriadExecutor(const Logger::Ptr& log);
     ~MyriadExecutor() = default;
-
-    /**
-     * @brief Get myriad device
-     * @return Already booted and empty device or new booted device
-     */
-    DevicePtr openDevice(std::vector<DevicePtr> &devicePool, const MyriadConfig& config);
-
-    static void closeDevices(std::vector<DevicePtr> &devicePool, std::shared_ptr<IMvnc> mvnc);
 
     void allocateGraph(DevicePtr &device,
                        GraphDesc &graphDesc,
@@ -105,40 +56,12 @@ public:
 
     void getResult(GraphDesc &graphDesc, void *result_data, unsigned int result_bytes);
 
-    static std::string ncStatusToStr(ncGraphHandle_t *graphHandle, ncStatus_t status);
-
     std::vector<float> getPerfTimeInfo(ncGraphHandle_t *graphHandle);
 
-    void printThrottlingStatus();
-
     static float GetThermal(const DevicePtr& device);
-
-    template<typename T>
-    static std::vector<T> getGraphInfo(
-            ncGraphHandle_t* graphHandle,
-            ncGraphOption_t graphOption,
-            int numElems) {
-        std::vector<T> out(numElems);
-
-        unsigned int infoByteSize = numElems * sizeof(T);
-        if (ncGraphGetOption(graphHandle, graphOption, out.data(), &infoByteSize) != NC_OK) {
-            out.clear();
-        }
-
-        return out;
-    }
-
-private:
-    /**
-     * @brief Try to boot any available device that suitable for selected platform and protocol
-     * @param configPlatform Boot the selected platform
-     * @param configProtocol Boot device with selected protocol
-     */
-    ncStatus_t bootNextDevice(std::vector<DevicePtr> &devicePool,
-                              const MyriadConfig& config);
 };
 
-typedef std::shared_ptr<MyriadExecutor> MyriadExecutorPtr;
+using MyriadExecutorPtr = std::shared_ptr<MyriadExecutor>;
 
 }  // namespace MyriadPlugin
 }  // namespace vpu
