+++ inference-engine/src/vpu/myriad_plugin/myriad_mvnc_wraper.cpp
@@ -43,3 +43,37 @@ std::vector<std::string> Mvnc::AvailableDevicesNames() const {
 
     return availableDevices;
 }
+
+std::string Mvnc::ncStatusToStr(ncGraphHandle_t *graphHandle, ncStatus_t status) {
+#define MVNC_STATUS_TO_STR(E) case E: return #E;
+    switch (status) {
+        MVNC_STATUS_TO_STR(NC_OK)
+        MVNC_STATUS_TO_STR(NC_BUSY)
+        MVNC_STATUS_TO_STR(NC_ERROR)
+        MVNC_STATUS_TO_STR(NC_OUT_OF_MEMORY)
+        MVNC_STATUS_TO_STR(NC_DEVICE_NOT_FOUND)
+        MVNC_STATUS_TO_STR(NC_INVALID_PARAMETERS)
+        MVNC_STATUS_TO_STR(NC_TIMEOUT)
+        MVNC_STATUS_TO_STR(NC_MVCMD_NOT_FOUND)
+        MVNC_STATUS_TO_STR(NC_NOT_ALLOCATED)
+        MVNC_STATUS_TO_STR(NC_UNAUTHORIZED)
+        MVNC_STATUS_TO_STR(NC_UNSUPPORTED_FEATURE)
+        MVNC_STATUS_TO_STR(NC_UNSUPPORTED_GRAPH_FILE)
+        MVNC_STATUS_TO_STR(NC_UNSUPPORTED_CONFIGURATION_FILE)
+        case NC_MYRIAD_ERROR: {
+            if (graphHandle == nullptr) {
+                return "NC_MYRIAD_ERROR";
+            } else {
+                auto debugInfo = getGraphInfo<char>(graphHandle, NC_RO_GRAPH_DEBUG_INFO, NC_DEBUG_BUFFER_SIZE);
+                if (debugInfo.empty()) {
+                    return "NC_MYRIAD_ERROR";
+                } else {
+                    return std::string(debugInfo.begin(), debugInfo.end());
+                }
+            }
+        }
+        default:
+            return "UNKNOWN MVNC STATUS";
+    }
+#undef MVNC_STATUS_TO_STR
+}
